+++ src/app/products/page.module.css
@@ -1,0 +1,45 @@
+/* Dark background for the table container */
+.tableContainer {
+    border-radius: 10px;
+    padding: 16px;
+  }
+  
+  /* Table Rows */
+  .tableRow {
+    background-color: #07101d;
+    padding: 12px;
+    border-radius: 8px;
+    margin-bottom: 8px;
+    transition: background 0.3s;
+  }
+  
+  .tableRow:hover {
+    background-color: #30363d;
+  }
+  
+  /* Profile Picture */
+  .profilePic {
+    width: 40px;
+    height: 40px;
+    background-color: #6a5acd;
+    border-radius: 50%;
+  }
+  
+  /* Gender Badge */
+  .genderBadge {
+    padding: 4px 10px;
+    border-radius: 12px;
+    font-size: 0.875rem;
+    font-weight: bold;
+    text-align: center;
+  }
+  
+  /* More Options Button */
+  .moreBtn {
+    background: none;
+    border: none;
+    font-size: 1.5rem;
+    color: white;
+    cursor: pointer;
+  }
+
